Extract reusable inference admission and scope lease cleanup
