+++ Sources/MereRunCLI/Support/RuntimeModelPool.swift
@@ -182,62 +182,6 @@ struct RuntimeCapabilityStatus: Codable, Equatable, Sendable {
     let detail: String
 }
 
-struct RuntimeRequestAdmissionSnapshot: Codable, Equatable, Sendable {
-    let maxActiveRequests: Int
-    let activeRequests: Int
-    let queuedRequests: Int
-    let totalAdmittedRequests: Int
-    let totalCompletedRequests: Int
-    let totalCancelledRequests: Int
-    let admissionPaused: Bool?
-    let pressure: String?
-    var activeRequestDetails: [RuntimeActiveRequestSnapshot]? = nil
-    var lastClientDisconnectAt: Date? = nil
-    var lastCancellationAt: Date? = nil
-    var lastSlotReleaseAt: Date? = nil
-
-    init(
-        maxActiveRequests: Int,
-        activeRequests: Int,
-        queuedRequests: Int,
-        totalAdmittedRequests: Int,
-        totalCompletedRequests: Int,
-        totalCancelledRequests: Int,
-        admissionPaused: Bool? = nil,
-        pressure: String? = nil,
-        activeRequestDetails: [RuntimeActiveRequestSnapshot]? = nil,
-        lastClientDisconnectAt: Date? = nil,
-        lastCancellationAt: Date? = nil,
-        lastSlotReleaseAt: Date? = nil
-    ) {
-        self.maxActiveRequests = maxActiveRequests
-        self.activeRequests = activeRequests
-        self.queuedRequests = queuedRequests
-        self.totalAdmittedRequests = totalAdmittedRequests
-        self.totalCompletedRequests = totalCompletedRequests
-        self.totalCancelledRequests = totalCancelledRequests
-        self.admissionPaused = admissionPaused
-        self.pressure = pressure
-        self.activeRequestDetails = activeRequestDetails
-        self.lastClientDisconnectAt = lastClientDisconnectAt
-        self.lastCancellationAt = lastCancellationAt
-        self.lastSlotReleaseAt = lastSlotReleaseAt
-    }
-}
-
-struct RuntimeActiveRequestSnapshot: Codable, Equatable, Sendable {
-    let id: UUID
-    let admittedAt: Date
-    let modelID: String?
-    let streaming: Bool?
-    let requestedMaxTokens: Int?
-    let toolCount: Int?
-    let phase: String
-    let phaseDetail: String?
-    let firstTokenAt: Date?
-    let generatedTokenUpdates: Int
-}
-
 struct RuntimeFutureStats: Codable, Equatable, Sendable {
     let available: Bool
     let detail: String
@@ -531,326 +475,6 @@ struct RuntimeMemorySnapshot: Codable, Equatable, Sendable {
         activeModelCount = try container.decode(Int.self, forKey: .activeModelCount)
         guardTier = try container.decodeIfPresent(RuntimeMemoryGuardTier.self, forKey: .guardTier) ?? .balanced
         pressure = try container.decode(String.self, forKey: .pressure)
-    }
-}
-
-enum RuntimeMemoryPressureLevel: String, Codable, Equatable, Sendable {
-    case disabled
-    case unknown
-    case nominal
-    case elevated
-    case critical
-}
-
-enum RuntimeMemoryGuardTier: String, Codable, CaseIterable, Equatable, Sendable {
-    case off
-    case safe
-    case balanced
-    case aggressive
-    case custom
-
-    static let `default` = RuntimeMemoryGuardTier.balanced
-}
-
-struct RuntimeMemorySample: Equatable, Sendable {
-    let physicalBytes: UInt64
-    let residentBytes: UInt64?
-    let physicalFootprintBytes: UInt64?
-    let availableBytes: UInt64?
-    let freeBytes: UInt64?
-    let activeBytes: UInt64?
-    let inactiveBytes: UInt64?
-
-    init(
-        physicalBytes: UInt64,
-        residentBytes: UInt64?,
-        physicalFootprintBytes: UInt64? = nil,
-        availableBytes: UInt64? = nil,
-        freeBytes: UInt64? = nil,
-        activeBytes: UInt64? = nil,
-        inactiveBytes: UInt64? = nil
-    ) {
-        self.physicalBytes = physicalBytes
-        self.residentBytes = residentBytes
-        self.physicalFootprintBytes = physicalFootprintBytes
-        self.availableBytes = availableBytes
-        self.freeBytes = freeBytes
-        self.activeBytes = activeBytes
-        self.inactiveBytes = inactiveBytes
-    }
-
-    static func current() -> RuntimeMemorySample {
-        RuntimeProcessMemory.currentSample()
-    }
-}
-
-struct RuntimeMemoryPressurePolicy: Equatable, Sendable {
-    let tier: RuntimeMemoryGuardTier
-    let customCeilingBytes: UInt64?
-    let softLimitFraction: Double
-    let hardLimitFraction: Double
-
-    static let `default` = RuntimeMemoryPressurePolicy(
-        tier: .default
-    )
-
-    init(
-        tier: RuntimeMemoryGuardTier = .default,
-        customCeilingBytes: UInt64? = nil,
-        softLimitFraction: Double = 0.90,
-        hardLimitFraction: Double = 0.95
-    ) {
-        self.tier = tier
-        self.customCeilingBytes = customCeilingBytes
-        self.softLimitFraction = softLimitFraction
-        self.hardLimitFraction = hardLimitFraction
-    }
-
-    func pressure(for sample: RuntimeMemorySample) -> RuntimeMemoryPressureLevel {
-        guard tier != .off else {
-            return .disabled
-        }
-        guard let currentBytes = currentBytes(for: sample),
-              let limits = limits(for: sample) else {
-            return .unknown
-        }
-        if currentBytes >= limits.hard {
-            return .critical
-        }
-        if currentBytes >= limits.soft {
-            return .elevated
-        }
-        return .nominal
-    }
-
-    func projectedPressure(
-        for sample: RuntimeMemorySample,
-        additionalBytes: UInt64
-    ) -> RuntimeMemoryPressureLevel {
-        guard tier != .off else {
-            return .disabled
-        }
-        guard let currentBytes = currentBytes(for: sample),
-              let limits = limits(for: sample) else {
-            return .unknown
-        }
-        let (projected, overflow) = currentBytes.addingReportingOverflow(additionalBytes)
-        guard !overflow else {
-            return .critical
-        }
-        if projected >= limits.hard {
-            return .critical
-        }
-        if projected >= limits.soft {
-            return .elevated
-        }
-        return .nominal
-    }
-
-    func currentBytes(for sample: RuntimeMemorySample) -> UInt64? {
-        sample.physicalFootprintBytes ?? sample.residentBytes
-    }
-
-    func limits(for sample: RuntimeMemorySample) -> (ceiling: UInt64, soft: UInt64, hard: UInt64)? {
-        guard tier != .off, sample.physicalBytes > 0 else {
-            return nil
-        }
-        let ceiling: UInt64
-        switch tier {
-        case .off:
-            return nil
-        case .custom:
-            let custom = customCeilingBytes ?? 0
-            guard custom > 0 else {
-                return nil
-            }
-            ceiling = min(custom, staticCeiling(for: sample))
-        case .safe, .balanced, .aggressive:
-            ceiling = min(staticCeiling(for: sample), dynamicCeiling(for: sample))
-        }
-        guard ceiling > 0 else {
-            return nil
-        }
-        let soft = UInt64((Double(ceiling) * softLimitFraction).rounded(.down))
-        let hard = UInt64((Double(ceiling) * hardLimitFraction).rounded(.down))
-        return (ceiling, soft, hard)
-    }
-
-    private func staticCeiling(for sample: RuntimeMemorySample) -> UInt64 {
-        let reserve = staticReserveBytes(forPhysicalBytes: sample.physicalBytes)
-        guard sample.physicalBytes > reserve else {
-            return 0
-        }
-        return sample.physicalBytes - reserve
-    }
-
-    private func dynamicCeiling(for sample: RuntimeMemorySample) -> UInt64 {
-        guard let currentBytes = currentBytes(for: sample) else {
-            return staticCeiling(for: sample)
-        }
-        if let freeBytes = sample.freeBytes,
-           let inactiveBytes = sample.inactiveBytes,
-           let activeBytes = sample.activeBytes {
-            return currentBytes
-                + freeBytes
-                + inactiveBytes
-                + UInt64((Double(activeBytes) * activeReclaimRatio).rounded(.down))
-        }
-        if let availableBytes = sample.availableBytes {
-            return currentBytes + availableBytes
-        }
-        return sample.physicalBytes
-    }
-
-    private var activeReclaimRatio: Double {
-        switch tier {
-        case .off, .custom:
-            return 0
-        case .safe:
-            return 0.20
-        case .balanced:
-            return 0.50
-        case .aggressive:
-            return 0.80
-        }
-    }
-
-    private func staticReserveBytes(forPhysicalBytes physicalBytes: UInt64) -> UInt64 {
-        let gib = UInt64(1024 * 1024 * 1024)
-        let smallSystemThreshold = 24 * gib
-        if physicalBytes < smallSystemThreshold {
-            return 4 * gib
-        }
-        switch tier {
-        case .off:
-            return physicalBytes
-        case .safe:
-            return 8 * gib
-        case .balanced:
-            return 6 * gib
-        case .aggressive:
-            return 4 * gib
-        case .custom:
-            return 2 * gib
-        }
-    }
-}
-
-private enum RuntimeProcessMemory {
-    static func currentSample() -> RuntimeMemorySample {
-        let physicalBytes = ProcessInfo.processInfo.physicalMemory
-        let residentBytes = currentResidentBytes()
-        let physicalFootprintBytes = currentPhysicalFootprintBytes()
-        let host = currentHostMemory()
-        return RuntimeMemorySample(
-            physicalBytes: physicalBytes,
-            residentBytes: residentBytes,
-            physicalFootprintBytes: physicalFootprintBytes,
-            availableBytes: host.availableBytes,
-            freeBytes: host.freeBytes,
-            activeBytes: host.activeBytes,
-            inactiveBytes: host.inactiveBytes
-        )
-    }
-
-    private static func currentPhysicalFootprintBytes() -> UInt64? {
-        #if canImport(Darwin)
-        var info = rusage_info_v4()
-        let result = withUnsafeMutablePointer(to: &info) { pointer in
-            // The Darwin module imports `rusage_info_t *` as a pointer to an
-            // optional raw pointer. Rebinding the struct storage mirrors the
-            // C API's required `(rusage_info_t *)&info` cast; passing `&raw`
-            // would instead let the kernel overwrite the pointer variable.
-            pointer.withMemoryRebound(
-                to: rusage_info_t?.self,
-                capacity: 1
-            ) { rebound in
-                proc_pid_rusage(getpid(), RUSAGE_INFO_V4, rebound)
-            }
-        }
-        guard result == 0 else {
-            return nil
-        }
-        return info.ri_phys_footprint
-        #else
-        return nil
-        #endif
-    }
-
-    private static func currentResidentBytes() -> UInt64? {
-        #if canImport(Darwin)
-        var info = mach_task_basic_info()
-        var count = mach_msg_type_number_t(MemoryLayout<mach_task_basic_info>.size) / 4
-        let result = withUnsafeMutablePointer(to: &info) { pointer in
-            pointer.withMemoryRebound(to: integer_t.self, capacity: Int(count)) { rebound in
-                task_info(mach_task_self_, task_flavor_t(MACH_TASK_BASIC_INFO), rebound, &count)
-            }
-        }
-        guard result == KERN_SUCCESS else {
-            return nil
-        }
-        return UInt64(info.resident_size)
-        #elseif os(Linux)
-        guard let status = try? String(contentsOfFile: "/proc/self/status") else {
-            return nil
-        }
-        for line in status.split(separator: "\n") where line.hasPrefix("VmRSS:") {
-            let parts = line.split(separator: " ").compactMap { UInt64($0) }
-            guard let kilobytes = parts.first else {
-                return nil
-            }
-            return kilobytes * 1024
-        }
-        return nil
-        #else
-        return nil
-        #endif
-    }
-
-    private static func currentHostMemory() -> (
-        availableBytes: UInt64?,
-        freeBytes: UInt64?,
-        activeBytes: UInt64?,
-        inactiveBytes: UInt64?
-    ) {
-        #if canImport(Darwin)
-        var stats = vm_statistics64()
-        var count = mach_msg_type_number_t(MemoryLayout<vm_statistics64>.size) / 4
-        let result = withUnsafeMutablePointer(to: &stats) { pointer in
-            pointer.withMemoryRebound(to: integer_t.self, capacity: Int(count)) { rebound in
-                host_statistics64(mach_host_self(), HOST_VM_INFO64, rebound, &count)
-            }
-        }
-        guard result == KERN_SUCCESS else {
-            return (nil, nil, nil, nil)
-        }
-        let pageSize = UInt64(getpagesize())
-        let free = UInt64(stats.free_count) * pageSize
-        let active = UInt64(stats.active_count) * pageSize
-        let inactive = UInt64(stats.inactive_count) * pageSize
-        return (free + inactive, free, active, inactive)
-        #elseif os(Linux)
-        guard let meminfo = try? String(contentsOfFile: "/proc/meminfo") else {
-            return (nil, nil, nil, nil)
-        }
-        var values: [String: UInt64] = [:]
-        for line in meminfo.split(separator: "\n") {
-            let parts = line.split(separator: " ")
-            guard let key = parts.first?.dropLast(),
-                  let kilobytes = parts.dropFirst().compactMap({ UInt64($0) }).first else {
-                continue
-            }
-            values[String(key)] = kilobytes * 1024
-        }
-        return (
-            values["MemAvailable"],
-            values["MemFree"],
-            values["Active"],
-            values["Inactive"]
-        )
-        #else
-        return (nil, nil, nil, nil)
-        #endif
     }
 }
 
@@ -2141,387 +1765,6 @@ actor RuntimeModelPool {
         states[id] = state
     }
     #endif
-}
-
-actor RuntimeRequestAdmission {
-    private struct Waiter {
-        let id: UUID
-        let continuation: CheckedContinuation<RuntimeRequestAdmissionLease, Error>
-    }
-
-    private enum WaiterState {
-        case pending
-        case waiting
-        case cancelled
-    }
-
-    private struct ActiveRequestState {
-        let id: UUID
-        let admittedAt: Date
-        var modelID: String?
-        var streaming: Bool?
-        var requestedMaxTokens: Int?
-        var toolCount: Int?
-        var phase = "admitted"
-        var phaseDetail: String?
-        var firstTokenAt: Date?
-        var generatedTokenUpdates = 0
-        var lastEventSequence: UInt64 = 0
-
-        var snapshot: RuntimeActiveRequestSnapshot {
-            RuntimeActiveRequestSnapshot(
-                id: id,
-                admittedAt: admittedAt,
-                modelID: modelID,
-                streaming: streaming,
-                requestedMaxTokens: requestedMaxTokens,
-                toolCount: toolCount,
-                phase: phase,
-                phaseDetail: phaseDetail,
-                firstTokenAt: firstTokenAt,
-                generatedTokenUpdates: generatedTokenUpdates
-            )
-        }
-    }
-
-    private let maxActiveRequests: Int
-    private let pressureProvider: @Sendable () async -> RuntimeMemoryPressureLevel
-    private var activeRequests = 0
-    private var waiters: [Waiter] = []
-    private var waiterStates: [UUID: WaiterState] = [:]
-    private var totalAdmittedRequests = 0
-    private var totalCompletedRequests = 0
-    private var totalCancelledRequests = 0
-    private var activeRequestStates: [UUID: ActiveRequestState] = [:]
-    private var lastClientDisconnectAt: Date?
-    private var lastCancellationAt: Date?
-    private var lastSlotReleaseAt: Date?
-
-    init(
-        maxActiveRequests: Int,
-        pressureProvider: @escaping @Sendable () async -> RuntimeMemoryPressureLevel = { .nominal }
-    ) {
-        precondition(maxActiveRequests > 0, "maxActiveRequests must be positive")
-        self.maxActiveRequests = maxActiveRequests
-        self.pressureProvider = pressureProvider
-    }
-
-    func acquire() async throws -> RuntimeRequestAdmissionLease {
-        await drain()
-        if await canAdmitNow(requireEmptyQueue: true) {
-            return admit()
-        }
-
-        let waiterID = UUID()
-        waiterStates[waiterID] = .pending
-        return try await withTaskCancellationHandler {
-            try await withCheckedThrowingContinuation { continuation in
-                enqueueWaiter(id: waiterID, continuation: continuation)
-            }
-        } onCancel: { [weak self] in
-            guard let self else { return }
-            Task {
-                await self.cancelWaiter(id: waiterID)
-            }
-        }
-    }
-
-    /// Acquires immediately without joining the FIFO, or returns nil when an
-    /// active/queued request already owns the capacity. Eviction uses this to
-    /// avoid waiting behind and then unloading a runtime that was active when
-    /// maintenance began.
-    func tryAcquire() async -> RuntimeRequestAdmissionLease? {
-        guard waiters.isEmpty, activeRequests < maxActiveRequests else {
-            return nil
-        }
-        let pressure = await pressureProvider()
-        guard waiters.isEmpty,
-              activeRequests < maxActiveRequests,
-              !admissionPaused(for: pressure) || activeRequests == 0 else {
-            return nil
-        }
-        return admit()
-    }
-
-    fileprivate func configureLease(
-        id: UUID,
-        modelID: String,
-        streaming: Bool,
-        requestedMaxTokens: Int,
-        toolCount: Int
-    ) {
-        guard var state = activeRequestStates[id] else { return }
-        state.modelID = modelID
-        state.streaming = streaming
-        state.requestedMaxTokens = requestedMaxTokens
-        state.toolCount = toolCount
-        activeRequestStates[id] = state
-    }
-
-    func recordProgress(
-        id: UUID,
-        sequence: UInt64,
-        progress: ChatProgress,
-        observedAt: Date = Date()
-    ) {
-        guard var state = activeRequestStates[id] else { return }
-        if progress.stage == .generating, progress.message?.isEmpty == false {
-            if let firstTokenAt = state.firstTokenAt {
-                state.firstTokenAt = min(firstTokenAt, observedAt)
-            } else {
-                state.firstTokenAt = observedAt
-            }
-            state.generatedTokenUpdates += 1
-        }
-        guard state.phase != "cancelling", sequence > state.lastEventSequence else {
-            activeRequestStates[id] = state
-            return
-        }
-        state.lastEventSequence = sequence
-        switch progress.stage {
-        case .loadingModel:
-            state.phase = "loading_model"
-            state.phaseDetail = progress.message
-        case .encoding:
-            state.phase = "prefill"
-            state.phaseDetail = progress.message
-        case .generating:
-            state.phase = "decode"
-            state.phaseDetail = nil
-        }
-        activeRequestStates[id] = state
-    }
-
-    func recordClientDisconnect(id: UUID, sequence: UInt64, observedAt: Date = Date()) {
-        guard var state = activeRequestStates[id] else { return }
-        state.phase = "cancelling"
-        state.phaseDetail = "Client disconnected"
-        state.lastEventSequence = max(state.lastEventSequence, sequence)
-        activeRequestStates[id] = state
-        lastClientDisconnectAt = observedAt
-    }
-
-    fileprivate func releaseLease(id: UUID, cancelled: Bool) async {
-        activeRequests = max(0, activeRequests - 1)
-        activeRequestStates.removeValue(forKey: id)
-        lastSlotReleaseAt = Date()
-        if cancelled {
-            totalCancelledRequests += 1
-            lastCancellationAt = Date()
-        } else {
-            totalCompletedRequests += 1
-        }
-        await drain()
-    }
-
-    func snapshot() async -> RuntimeRequestAdmissionSnapshot {
-        let pressure = await pressureProvider()
-        return RuntimeRequestAdmissionSnapshot(
-            maxActiveRequests: maxActiveRequests,
-            activeRequests: activeRequests,
-            queuedRequests: waiters.count,
-            totalAdmittedRequests: totalAdmittedRequests,
-            totalCompletedRequests: totalCompletedRequests,
-            totalCancelledRequests: totalCancelledRequests,
-            admissionPaused: admissionPaused(for: pressure),
-            pressure: pressure.rawValue,
-            activeRequestDetails: activeRequestStates.values
-                .map(\.snapshot)
-                .sorted { $0.admittedAt < $1.admittedAt },
-            lastClientDisconnectAt: lastClientDisconnectAt,
-            lastCancellationAt: lastCancellationAt,
-            lastSlotReleaseAt: lastSlotReleaseAt
-        )
-    }
-
-    private func enqueueWaiter(
-        id: UUID,
-        continuation: CheckedContinuation<RuntimeRequestAdmissionLease, Error>
-    ) {
-        if waiterStates[id] == .cancelled {
-            waiterStates.removeValue(forKey: id)
-            continuation.resume(throwing: CancellationError())
-            return
-        }
-
-        waiterStates[id] = .waiting
-        waiters.append(Waiter(id: id, continuation: continuation))
-    }
-
-    private func cancelWaiter(id: UUID) {
-        switch waiterStates[id] {
-        case .pending:
-            waiterStates[id] = .cancelled
-            totalCancelledRequests += 1
-        case .waiting:
-            guard let index = waiters.firstIndex(where: { $0.id == id }) else {
-                waiterStates.removeValue(forKey: id)
-                return
-            }
-            let waiter = waiters.remove(at: index)
-            waiterStates.removeValue(forKey: id)
-            totalCancelledRequests += 1
-            waiter.continuation.resume(throwing: CancellationError())
-        case .cancelled, nil:
-            return
-        }
-    }
-
-    private func drain() async {
-        while activeRequests < maxActiveRequests, !waiters.isEmpty {
-            guard await canAdmitNow() else {
-                return
-            }
-            // Pressure sampling yields the actor. A concurrent drain may have
-            // consumed capacity or cancellation may have emptied the FIFO, so
-            // revalidate both before removing its first waiter.
-            guard activeRequests < maxActiveRequests, !waiters.isEmpty else {
-                return
-            }
-            let waiter = waiters.removeFirst()
-            guard waiterStates[waiter.id] != .cancelled else {
-                waiterStates.removeValue(forKey: waiter.id)
-                waiter.continuation.resume(throwing: CancellationError())
-                continue
-            }
-            waiterStates.removeValue(forKey: waiter.id)
-            waiter.continuation.resume(returning: admit())
-        }
-    }
-
-    private func admit() -> RuntimeRequestAdmissionLease {
-        let id = UUID()
-        activeRequests += 1
-        totalAdmittedRequests += 1
-        activeRequestStates[id] = ActiveRequestState(id: id, admittedAt: Date())
-        return RuntimeRequestAdmissionLease(id: id, admission: self)
-    }
-
-    private func canAdmitNow(requireEmptyQueue: Bool = false) async -> Bool {
-        guard activeRequests < maxActiveRequests,
-              !requireEmptyQueue || waiters.isEmpty else {
-            return false
-        }
-        let pressure = await pressureProvider()
-        // `pressureProvider` is an actor reentrancy point. Never rely on the
-        // capacity/FIFO snapshot taken before it suspended.
-        guard activeRequests < maxActiveRequests,
-              !requireEmptyQueue || waiters.isEmpty else {
-            return false
-        }
-        guard admissionPaused(for: pressure) else {
-            return true
-        }
-        return activeRequests == 0
-    }
-
-    private func admissionPaused(for pressure: RuntimeMemoryPressureLevel) -> Bool {
-        switch pressure {
-        case .elevated, .critical:
-            return true
-        case .disabled, .unknown, .nominal:
-            return false
-        }
-    }
-}
-
-final class RuntimeRequestAdmissionLease: @unchecked Sendable {
-    private let id: UUID
-    private let admission: RuntimeRequestAdmission
-    private let lock = NSLock()
-    private var released = false
-    private var clientDisconnected = false
-    private var nextEventSequence: UInt64 = 0
-
-    init(id: UUID, admission: RuntimeRequestAdmission) {
-        self.id = id
-        self.admission = admission
-    }
-
-    var requestID: UUID { id }
-
-    deinit {
-        guard markReleased() else { return }
-        let admission = admission
-        let id = id
-        Task {
-            await admission.releaseLease(id: id, cancelled: false)
-        }
-    }
-
-    func configure(
-        modelID: String,
-        streaming: Bool,
-        requestedMaxTokens: Int,
-        toolCount: Int
-    ) async {
-        await admission.configureLease(
-            id: id,
-            modelID: modelID,
-            streaming: streaming,
-            requestedMaxTokens: requestedMaxTokens,
-            toolCount: toolCount
-        )
-    }
-
-    func observe(_ progress: ChatProgress) {
-        guard let sequence = reserveEventSequence() else { return }
-        let admission = admission
-        let id = id
-        let observedAt = Date()
-        Task {
-            await admission.recordProgress(
-                id: id,
-                sequence: sequence,
-                progress: progress,
-                observedAt: observedAt
-            )
-        }
-    }
-
-    func observeClientDisconnect() {
-        guard let sequence = reserveDisconnectSequence() else { return }
-        let admission = admission
-        let id = id
-        let observedAt = Date()
-        Task {
-            await admission.recordClientDisconnect(
-                id: id,
-                sequence: sequence,
-                observedAt: observedAt
-            )
-        }
-    }
-
-    func release(cancelled: Bool = false) async {
-        guard markReleased() else { return }
-        await admission.releaseLease(id: id, cancelled: cancelled)
-    }
-
-    private func markReleased() -> Bool {
-        lock.lock()
-        defer { lock.unlock() }
-        guard !released else { return false }
-        released = true
-        return true
-    }
-
-    private func reserveEventSequence() -> UInt64? {
-        lock.lock()
-        defer { lock.unlock() }
-        guard !released, !clientDisconnected else { return nil }
-        nextEventSequence += 1
-        return nextEventSequence
-    }
-
-    private func reserveDisconnectSequence() -> UInt64? {
-        lock.lock()
-        defer { lock.unlock() }
-        guard !released, !clientDisconnected else { return nil }
-        clientDisconnected = true
-        nextEventSequence += 1
-        return nextEventSequence
-    }
 }
 
 final class RuntimeModelLease: @unchecked Sendable {
